Renamed finite_element_test.h to .hpp.
